plugins: add PrivateGithub plugin
which translates requests against configurable github.com/... URLs
to api.github.com requests.
It uses an environment provided TOKEN, so no secret information
will be exposed to any potentially persistent file.
The plugin also comes with a configurable option to deal with
potential rate limits of the github API.

With the help of this plugin it is possible to check links in
private github repositories.

Relates to #404

Signed-off-by: Konrad Weihmann <konrad.weihmann@avnet.eu>